requirements/base.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
asgiref==3.6.0
attrs==22.2.0
Babel==2.11.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.1.5
django-address==0.2.8
django-filter==22.1
django-heroku==0.3.1
django-import-export==3.0.2
django-money==3.0.0
django-phonenumber-field==7.0.2
django-qr-code==3.1.1
django-tinymce==3.5.0
django-uuidfield==0.5.0
django-versatileimagefield==2.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
exceptiongroup==1.1.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
Markdown==3.4.1
MarkupPy==1.14
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.0
pathspec==0.10.3
phonenumbers==8.13.4
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.1
py-moneyed==2.0
pydantic==1.10.4
PyJWT==2.6.0
pytest==7.2.1
pytest-django==4.5.2
pytest-django-factories==0.2.3
pytest-django-filefield==0.1.2
python-decouple==3.4
python-magic==0.4.27
pytz==2021.1
PyYAML==6.0
requests==2.28.2
rest-framework-money-field==2.0.0
segno==1.5.2
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.11.0
